backend/requirements.txt
Flask==3.0.0
gunicorn==23.0.0
pydantic[email]==2.5.0
python-dotenv==1.0.0
PyJWT==2.8.0
psycopg2-binary==2.9.9
redis==5.0.1
requests==2.31.0
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
'pydantic[email]'
